refactor(researcher): Mover interfaces y entidades al dominio
- Movidas clases al dominio:
  - QueryValidation a entities/query_validation.py
  - SearchEngine a entities/search_engine.py
  - ResearchRepository a repositories/research_repository.py
- Actualizadas importaciones en researcher.py
- Sin cambios funcionales
