+++ app/agents/researcher/domain/entities/__init__.py
@@ -1,3 +1,5 @@
 from .metrics_data import MetricsData
+from .search_engine import SearchEngine
+from .query_validation import QueryValidation
 
-__all__ = ['MetricsData']
+__all__ = ['MetricsData', 'SearchEngine', 'QueryValidation']
