+++ app/agents/researcher/domain/entities/search_engine.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+class SearchEngine(Enum):
+    TAVILY = "tavily"
+    GEMINI = "gemini"
+    DEEP_RESEARCH = "deep_research"
