+++ app/agents/researcher/domain/repositories/__init__.py
@@ -1,0 +1,3 @@
+from .research_repository import ResearchRepository
+
+__all__ = ['ResearchRepository']
